billing/requirements.txt
fastapi[standard]==0.115.6
sqlmodel==0.0.22
python-dotenv==1.0.1
PyJWT==2.10.1
pytest==8.3.4
cryptography